requirements.txt
Flask ==1.1.2
numpy ==1.20.1
vosk ==0.3.30
wave ==0.0.2
Morfessor ==2.0.6
Werkzeug ==1.0.1
pandas ==1.2.4
python-dotenv==0.18.0
Jinja2==2.10.3
gunicorn==20.0.4
MarkupSafe==1.1.1
itsdangerous==1.1.0
Click==7.1.2
nltk==3.6.1